requirements.txt
aerich==0.3.3
aiohttp==3.7.3
aiohttp-autoreload==0.0.1
aiohttp-cors==0.7.0
aiohttp-devtools==0.13.1
aiohttp-jinja2==1.2.0
aiohttp-swagger==1.0.14
aiohttp-swagger3==0.5.0
aiomysql==0.0.20
aioredis==1.3.1
aiosmtplib==1.0.5
aiosqlite==0.16.0
alembic==1.4.3
aliyun-python-sdk-core==2.13.29
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.12.0
aliyun-python-sdk-sts==3.0.2
appdirs==1.4.3
APScheduler==3.6.3
async-timeout==3.0.1
asyncoss==0.0.4
attrs==20.3.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
contextvars==2.4
crcmod==1.7
cryptography==3.2.1
devtools==0.6.1
fake-useragent==0.1.11
fastjsonschema==2.14.5
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
hiredis==1.1.0
idna==2.8
idna-ssl==1.1.0
immutables==0.11
iso8601==0.1.13
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
Mako==1.1.3
MarkupSafe==1.1.1
motor==2.0.0
multidict==4.5.2
oss2==2.13.1
pycparser==2.20
pycryptodome==3.9.9
pydantic==1.7.2
pyee==6.0.0
Pygments==2.7.2
pymongo==3.7.2
PyMySQL==0.9.2
PyPika==0.44.0
pyppeteer==0.0.25
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
selenium==3.141.0
six==1.13.0
SQLAlchemy==1.3.20
strict-rfc3339==0.7
tortoise-orm==0.16.18
tqdm==4.41.1
trafaret==1.2.0
trafaret-config==2.0.2
typing==3.7.4.3
typing-extensions==3.7.4.3
tzlocal==2.0.0
urllib3==1.24.1
watchgod==0.6
websockets==6.0
Werkzeug==1.0.1
yarl==1.3.0
